backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
